Refactor PresetsComponent
